Sam Fix mac build (#23)
* Changes to allow the sample viewer to be build on Mac

* Update QueryLocation.h

+++ sample_project/Source/sample_project/Geocoding/QueryLocation.h
@@ -56,10 +56,10 @@ private:
 	bool bShouldUpdateElevation = false;
 	FString Address;
 	APawn* PawnActor;
-	UINT16 StableFramesCounter; // Counting the frames during which the raycast has returned the same hit result
-	UINT16 FramesToWaitForLoading = 30; // Threshold for comparing the StableFramesCounter against
-	UINT16 RaycastCounter; // Counting the total number of raycasts performed for this location
-	UINT16 MaxRaycastAttemts = 200; // Threshold for comparing the RaycastCounter against
+	int StableFramesCounter; // Counting the frames during which the raycast has returned the same hit result
+	int FramesToWaitForLoading = 30; // Threshold for comparing the StableFramesCounter against
+	int RaycastCounter; // Counting the total number of raycasts performed for this location
+	int MaxRaycastAttemts = 200; // Threshold for comparing the RaycastCounter against
 	UStaticMesh* PinMesh;
 	UStaticMesh* PointMesh;
 	UMaterial* PointMaterial;
